--- gridcapa-task-manager-app/src/main/java/com/rte_france/farao/gridcapa_task_manager_app/TaskManagerApplication.java
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.rte_france.farao.gridcapa_task_manager_app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-
-/**
- * @author Sebastien Murgey {@literal <sebastien.murgey at rte-france.com>}
- */
-@SuppressWarnings("hideutilityclassconstructor")
-@SpringBootApplication
-@EnableConfigurationProperties(TaskManagerConfigurationProperties.class)
-@EntityScan(basePackages = "api.entities")
-public class TaskManagerApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(TaskManagerApplication.class, args);
-    }
-}
