--- gridcapa-task-manager-app/src/test/java/com/rte_france/farao/gridcapa_task_manager_app/TaskManagerControllerTest.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.rte_france.farao.gridcapa_task_manager_app;
-
-import com.rte_france.farao.gridcapa_task_manager_api.entities.Task;
-import com.rte_france.farao.gridcapa_task_manager_api.entities.TaskDto;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-import java.time.LocalDateTime;
-import java.util.Collections;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-/**
- * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
- */
-@SpringBootTest
-class TaskManagerControllerTest {
-
-    @MockBean
-    private TaskRepository taskRepository;
-
-    @Autowired
-    private TaskManagerController taskManagerController;
-
-    @Test
-    void testGetTaskOk() {
-        LocalDateTime taskTimestamp = LocalDateTime.parse("2021-09-30T23:00");
-        Task task = new Task(taskTimestamp, Collections.emptyList());
-        Mockito.when(taskRepository.findByTimestamp(taskTimestamp)).thenReturn(Optional.of(task));
-        ResponseEntity<TaskDto> taskResponse = taskManagerController.getTaskFromTimestamp(taskTimestamp.toString());
-
-        assertEquals(HttpStatus.OK, taskResponse.getStatusCode());
-    }
-}
